backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
PyPDF2==3.0.1
PyMuPDF==1.23.8
Pillow==10.1.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
cors==1.0.1
requests==2.31.0
pytesseract==0.3.10
reportlab==4.0.7
cryptography==41.0.7
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2